+++ src/main/kotlin/racingcar/view/ResultView.kt
@@ -1,0 +1,9 @@
+package racingcar.view
+
+object ResultView {
+    private const val RESULT = "실행 결과"
+
+    fun printResultTitle() {
+        println(RESULT)
+    }
+}
